Can we please have some consistancy
Using sprites to speed up loading time is lovely, but we have been left
with an inconsistant skin. If you are going to move the right shower into
a sprite, you move the left shower into a sprite. You don't move one and
leave the other. Also, if you change the right shower to have thicker
lines, you change the left shower to have thicker lines, you don't leave
them with different thicknesses of lines.

I would prefer to have a page that loads slightly slower but is consistant
than a page that loads quickly but is terribly inconsistant.