More Internet Explorer 8 fixes (Brickimedia/brickimedia#224)
* correct capitalization of the page name in the left sidebar -- IE8 doesn't understand the :first-of-type selector
* CSS gradient -- it's not as pretty as on more modern browsers (that understand CSS3), but it's something
* search suggestions' positioning fixed

Tested *only* on IE8, hence why the separate file is called ie8.css and it's loaded only for IE8. If it's deemed necessary and wanted to load these fixes for other IEs, we can change the rule (and filename) in Refreshed.skin.php accordingly.

Tested on: Windows XP Professional / IE8 / 1400x1050px