Less hacky notifications item, W3C fixes
* Changing notifications item in personal tools dropdown per
https://github.com/Brickimedia/Refreshed/commit/00ce3a3ff28e5707fce2f207
d627679638714493#commitcomment-9119517
* PHP fixes in accordance with W3C standards. As of this commit the
only errors on normal pages should be a rel=“archives” issue (beyond
our control AFAIK) and images without alt text. That said, we might
want to change the name of the search element in the header so
Special:Search can be validated, since the search bars in the header
and on the page both have the id #search.
* Deleting some unnecessary code in Refreshed.js (I thought I deleted
it before I pushed the last commit but apparently not)