More HTML5 semantic elements, reenable mobile logo
* More HTML5 semantic elements/HTML cleanup per
Brickimedia/brickimedia#420
* Reenabling the mobile logo. It appears in small mode and is defined
in LocalSettings.php as $wgRefreshedHeader[‘mobileimg’] (an image
element similar to $wgRefreshedHeader[‘img’]). Before it was disabled
because it broke the site if $wgRefreshedHeader[‘mobileimg’] was not
defined, but now Refreshed.skin.php checks if the logo is defined
before adding it so the page shouldn’t break.