Update WikiFont, lots of header cleanup
First steps of https://github.com/Brickimedia/brickimedia/issues/369

* Updating WikiFont to latest commit.
* When I tested the latest version of WikiFont on my local clone, it
broke a bunch of header CSS. As it turns out, that's because the CSS
was really sloppy. (My fault.) I removed and rewrote a lot of it and it
should be a lot cleaner now. To add to that, now the header height can
be adjusted just by changing two CSS properties, while it was a whole
mess to change it before. One issue that still remains is that images
are vertically centered using a margin-bottom that's a magic number. If
someone could come up with a better solution that'd be excellent.

Now that everything's been cleaned up it should be fairly
straightforward to add in the page action icons.